OCR/requirements.txt
# 核心依赖
opencv-python>=4.5.5
Pillow>=9.0.0
numpy>=1.20.0
paddleocr>=2.6.1.3
streamlit>=1.10.0
pandas>=1.3.0
python-docx>=0.8.11

# PaddleOCR依赖
paddlepaddle>=2.3.0

# 可选依赖（Tesseract支持）
pytesseract>=0.3.8